Unify `resolve` calls
