+++ packages/graph/hash_graph/lib/graph/src/store/postgres/knowledge/link/mod.rs
@@ -1,4 +1,5 @@
 mod read;
+mod resolve;
 
 use async_trait::async_trait;
 use error_stack::{Result, ResultExt};
