+++ packages/graph/hash_graph/lib/graph/src/store/postgres/ontology/data_type/mod.rs
@@ -1,4 +1,4 @@
-pub mod read;
+pub mod resolve;
 
 use async_trait::async_trait;
 use error_stack::{IntoReport, Result, ResultExt};
