+++ packages/graph/hash_graph/lib/graph/src/store/postgres/ontology/link_type/mod.rs
@@ -1,4 +1,4 @@
-mod read;
+mod resolve;
 
 use async_trait::async_trait;
 use error_stack::{IntoReport, Result, ResultExt};
