--- packages/graph/hash_graph/lib/graph/src/store/postgres/resolve/mod.rs
@@ -1,8 +1,0 @@
-mod context;
-mod entity;
-mod links;
-mod ontology;
-
-pub use self::{
-    context::PostgresContext, entity::EntityRecord, links::LinkRecord, ontology::OntologyRecord,
-};
